backend/matchmaking/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg==3.2.3
psycopg2-binary==2.9.10
PyJWT==2.9.0
redis==5.2.0
sqlparse==0.5.2
typing_extensions==4.12.2
dj-database-url==2.3.0
python-dateutil==2.8.2
